Completed JS to load fleet names and truck fleetnum values when drawing the operation table. Removed PHP init draw of table and shifted all drawing functionality of HTML table for the operations for this process to JS.
Newly added remove_process_data.php script added to remove operation entries in the table list by using a JS function whenn clicking a link for any of the rows in the table. JS completed to send a POST request to the php script and get the result. PHP script needs to be completed as it is not successfully deleting the record.